chore: moved Card to component folder
